django-keycloak-demo/requirements.txt
cffi                           1.14.1
chardet                        3.0.4
cryptography                   3.0
Django                         2.1.15
django-keycloak-auth           0.9.1
django-rest-framework-keycloak 0.2.0
djangorestframework            3.11.0
djangorestframework-jwt        1.11.0
ecdsa                          0.15
future                         0.18.2
httmock                        1.2.5
idna                           2.6
install                        1.3.3
mysqlclient                    2.0.1
oauthlib                       3.1.0
pip                            20.1.1
pyasn1                         0.4.8
pycparser                      2.20
pycryptodome                   3.9.8
pycryptosat                    0.6.1.1
PyJWT                          1.7.1
python-jose                    3.1.0
python-keycloak                0.21.0
pytz                           2020.1
requests                       2.24.0
requests-oauthlib              0.7.0
rsa                            4.6
setuptools                     46.4.0
six                            1.15.0
urllib3                        1.22
wheel                          0.34.2
